LPS-60923 Fix compile, move related classes to portal-service
