+++ portal-impl/src/META-INF/workflow-spring.xml
@@ -20,7 +20,7 @@
 			</bean>
 		</constructor-arg>
 		<constructor-arg>
-			<bean class="com.liferay.portal.messaging.proxy.MessagingProxyInvocationHandler" factory-method="getInvocationHandlerFactory" />
+			<bean class="com.liferay.portal.kernel.messaging.proxy.MessagingProxyInvocationHandler" factory-method="getInvocationHandlerFactory" />
 		</constructor-arg>
 	</bean>
 	<bean id="com.liferay.portal.workflow.WorkflowLinkAdvice" class="com.liferay.portal.workflow.WorkflowLinkAdvice" />
