+++ portal-impl/src/com/liferay/portal/messaging/proxy/MultiDestinationMessagingProxyInvocationHandler.java
@@ -17,7 +17,7 @@ package com.liferay.portal.messaging.proxy;
 import com.liferay.portal.kernel.messaging.proxy.BaseMultiDestinationProxyBean;
 import com.liferay.portal.kernel.messaging.proxy.ProxyModeThreadLocal;
 import com.liferay.portal.kernel.messaging.proxy.ProxyRequest;
-import com.liferay.portal.spring.aop.InvocationHandlerFactory;
+import com.liferay.portal.kernel.spring.aop.InvocationHandlerFactory;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
